+++ tests/src/result.tests.cpp
@@ -12,6 +12,8 @@ EXT_DISABLE_WARNINGS_POP
 
 #include <ext/result.hpp>
 
+#include <ext/utility_classes.hpp>
+
 using namespace ext;
 
 TEST(ResultTests, Construct)
@@ -667,5 +669,27 @@ TEST(ResultTests, MapOperator)
         const auto r2 = r1 | fn1 | fn2;
         ASSERT_TRUE(r2.success());
         ASSERT_EQ(r2.value(), 2U);
+    }
+}
+
+TEST(ResultTests, GetValueWhenResultIsRValue)
+{
+    using namespace ext::testing;
+
+    {
+        ClassHasNoCopyCtorButMovable object {};
+        Result<ClassHasNoCopyCtorButMovable, int> res { Success<ClassHasNoCopyCtorButMovable>(std::move(object)) };
+        object = std::move(res).value();
+    }
+}
+
+TEST(ResultTests, GetErrorWhenResultIsRValue)
+{
+    using namespace ext::testing;
+
+    {
+        ClassHasNoCopyCtorButMovable object {};
+        Result<int, ClassHasNoCopyCtorButMovable> res { Failure<ClassHasNoCopyCtorButMovable>(std::move(object)) };
+        object = std::move(res).error();
     }
 }